requirements.txt
Flask
flask-swagger-ui
flask-cors
azure-ai-vision-imageanalysis
python-decouple